requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Babel==4.0.0
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
stripe==10.5.0
requests==2.32.3
python-slugify==8.0.4
marshmallow==3.21.1
Flask-Limiter==3.6.0
